requirements.txt
playwright
schedule
python-dotenv
moviepy
openai
requests
fastapi
uvicorn
openai-whisper
gspread
google-auth
